Airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-microsoft-azure
apache-airflow-providers-databricks
flask-appbuilder
apache-airflow-providers-fab